requirements.txt
async-generator==1.10
attrs==20.3.0
autopep8==1.5.5
fuse-python==1.0.3
idna==3.1
outcome==1.1.0
pycodestyle==2.6.0
pydantic==1.8.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
sniffio==1.2.0
sortedcontainers==2.3.0
toml==0.10.2
trio==0.18.0
typing-extensions==3.7.4.3
watchdog==2.0.0
